Fixup compile warnings
Reviewed-on: http://review.couchbase.org/12585
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>

+++ tools/cbc.cc
@@ -346,7 +346,7 @@ static bool spool(string &data) {
     stringstream ss;
     char buffer[1024];
     size_t nr;
-    while ((nr = fread(buffer, 1, sizeof(buffer), stdin)) != -1) {
+    while ((nr = fread(buffer, 1, sizeof(buffer), stdin)) != (size_t)-1) {
         if (nr == 0) {
             break;
         }
